requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-pptx
markdown
beautifulsoup4
pillow
